Added auth request on FavoritesScreen

+++ shared/src/commonMain/kotlin/kfu/itis/freshnews/feature/auth/domain/usecase/GetUserIdUseCase.kt
@@ -2,5 +2,5 @@ package kfu.itis.freshnews.feature.auth.domain.usecase
 
 interface GetUserIdUseCase {
 
-    suspend operator fun invoke(): Int?
+    suspend operator fun invoke(): Long?
 }
