+++ shared/src/commonMain/kotlin/kfu/itis/freshnews/feature/auth/domain/usecase/SaveUserIdUseCase.kt
@@ -2,5 +2,5 @@ package kfu.itis.freshnews.feature.auth.domain.usecase
 
 interface SaveUserIdUseCase {
 
-    suspend operator fun invoke(id: Int)
+    suspend operator fun invoke(id: Long)
 }
